// modules/components/helpers/constants.d.ts
export declare const PLACEHOLDER = "+ Tag";
export declare const SECONDARY_PLACEHOLDER = "Enter a new tag";
export declare const KEYDOWN = "keydown";
export declare const KEYUP = "keyup";
export declare const FOCUS = "focus";
export declare const MAX_ITEMS_WARNING = "The number of items specified was greater than the property max-items.";
export declare const ACTIONS_KEYS: {
    DELETE: string;
    SWITCH_PREV: string;
    SWITCH_NEXT: string;
    TAB: string;
};
export declare const KEY_PRESS_ACTIONS: {
    8: string;
    37: string;
    39: string;
    9: string;
};